requirements.txt
crewai>=0.30.0
langchain>=0.2.0
langchain-openai>=0.1.0
langsmith>=0.1.0
instructor>=1.2.0
openai>=1.30.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
pytest>=8.0.0
